move foreground waiting to sigchld handler
prior to this commit, foreground jobs were reaped in the main loop and
background jobs were reaped in the sigchld handler

recall that signals of the same type do not queue. thus we must potentially
reap multiple processes in the sigchld handler. this could potentially reap
the fg process.

reap the foreground process in the sigchld handler. spin in the shell until
the foreground process is reaped